Highlight builtin names.
I want to use a variable to hold the Fish builtin keywords, because it's
a big list, and I don't want to manage an optimized regex by hand.

To use a variable in fish-font-lock-keywords-1 I had to use a backquote
and a comma to get the variable to be evaluated within the list passed
to font-lock-defaults.  Without it, font-lock-defaults fails with a
message about, "fish-builtin-commands-re" being an invalid function.