Fix infinite loop after upgrading to Fish 3.x
Something happened to Fish's variable scoping. This broke the
__find_hg_root function.  It could no longer walk up the directory
tree. The local variable `$dir` was no longer being decremented during
each loop. Instead, the variable `$dir` was unchanged on each iteration
of the while loop.

This is another fix which could have worked.
https://github.com/oh-my-fish/plugin-vcs/pull/7

In the end, I think this new implementation reads a little nicer, and it
delegates to a Fish builtin function that traverses up directories for
me. Now I can stop maintaining that little part of my config.