bash: remove aliases for switching jvm versions
These were useful back in the java 6, 7, and 8 days. But, now Java is
moving faster, and I don't use it as much. Plus these emitted errors
when opening new terminal tabs when I didn't have that version of Java
installed on my Mac.

In the future, prefer to use a proper version manager for Java like
jabba.