Reformat gitconfig
When using "git config" to set values it inserts a tab, which I don't
want in there, since I prefer spaces.

Emacs conf-mode seems to prefer 4 spaces for indentation. So instead of
fighting or configuring it explicitly, just give in and reformat the
file once to make it straightforward to edit going forward.