Remove dead undo-tree config
Doom Emacs uses the library emacs-undo-fu which I don't think has an
option to visualize the tree of undo/redo steps.

On the bright side, emacs-undo-fu doesn't have the problem of corrupting
emacs undo information. Apparently I've been living with that bug in
undo-tree for a long time now, and didn't know it.
http://ergoemacs.org/emacs/emacs_best_redo_mode.html