Spacemacs: configure evil to not put junk in my clipboard
The default settings on macOS create a situation where just highlighting
text in visual mode sends it to the clipboard. Which fills it with
partial junk and stuff I don't want.

Now to add text to the clipboard I need to highlight in in visual mode
and explicitly put it in the clipboard via Cmd+C.

Source: https://github.com/syl20bnr/spacemacs/issues/1504