spacemacs: fix key binding for markdown preview
The previous keybinding was global, which is not what I wanted. The new
keybinding is more idiomatic for Spacemacs. It uses the "o" namespace
reserved for user customization.  But it's within the context of major
mode shortcuts for the open buffer.

So, the mnemonic ends up being: "SPC(Leader) m(MajorMode) o(UserSpace)
p(Preview)".