Spacemacs: add function to delete multiple blank lines
Since nobody at Amazon uses CodeCommit, when you copy source code from
its web interface, what ends up in your clipboard is littered with
multiple excess blank lines.

I thought the Emacs function, delete-blank-lines, would do this, but it
is flagrantly false advertising and completely useless at this function.

So add a new function for this and add a keyboard shortcut for it.

Thanks to the comments in Mastering Emacs for this idea.