Spacemacs: port over my vim keybinding to toggle highlighting
Appease my muscle memory for how to toggle search highlighting on and
off. It turns out I've used that F2 key mapping for ":nohls" for so long
I miss it when it's gone.

Tested this solution and it works great.  I may not need all that
belt-and-suspenders stuff for defining the key binding. Polish that up
later.

Credit:
https://github.com/syl20bnr/spacemacs/issues/2719#issuecomment-310917138