Spacemacs: install packages in user-config
I was using use-package to install and configure some dependencies in my
user-config function.

When I recently added the "helpful" package to my config, I was getting
errors on startup that the "helpful" package could not be found.

Since I do not have `(setq use-package-always-ensure t)` in my
configuration anywhere, I guess I need to include the `:ensure t`
argument to the (use-package) macro to ensure the package is installed
when the init file is loaded.

This fixed the startup errors.