Start SSH agents in Fish shell
This function starts a new SSH agent if one doesn't exist already.
It is idempotent and can be called multiple times without starting
unnecessary agents.  Its saves its state to a file so all terminal
sessions can reuse the same running agent.

Be careful not to emit text in a non-interactive shell context. That
wreaks havoc with programs like scp and Emacs when running shell
commands in the background. So, if the shell isn't interactive, bail out
immediately.  Confirmed this worked with Emacs and Projectile.